scan.cc&friends: Increase variable sizes from unsigned int to size_t to allow for more than 179 mill points per scan on 64 bit architectures
The unsigned int type is 4 byte wide on all platforms. This means that it can
only store up to about 4.2 billion. This means, that when using this type to
address memory made of 3-tuples of doubles (which are 8 byte on all platforms),
then we are down to only 179 million possible 3-tuples because 4.2 billion / 8
/ 3 == 179 million.

Prior to this commit, if any scan exceeded 179 million points, a segmentation
fault would occur.  Thus, this commit changes all relevant size types from
unsigned int to size_t which is a 8 byte wide type on 64 bit platforms. This
allows up to 18 quintillion points per scan on architectures with an 8 byte
wide size_t.

To give a reasonable error message (instead of just a segmentation fault) on
limited 32bit architectures, this commit also adds a check before any
multiplication of point vector sizes to make sure that the result is still
within the bounds of size_t. This check is only be done on platforms where
size_t is 4 bytes long because we assume that nobody will ever need more than
18 quintillion points per scan (famous last words). Doing this check will not
add significant overhead on 32 bit platforms because it is not done in any
inner loop. On 64bit platforms this check will not impact runtime at all
because the if-block will be optimized away by the compiler with optimizations
are switched on.

While we are at it we also fix the existing unsigned int limit for frames and
bump that to size_t as well.

git-svn-id: svn://svn.code.sf.net/p/slam6d/code/trunk@1095 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ src/slam6d/managedScan.cc
@@ -248,7 +248,7 @@ void ManagedScan::get(unsigned int types)
   m_shared_scan->prefetch(types);
 }
 
-DataPointer ManagedScan::create(const std::string& identifier, unsigned int size)
+DataPointer ManagedScan::create(const std::string& identifier, size_t size)
 {
   // map identifiers to functions in SharedScan and scale back size
   // from bytes to number of points
@@ -341,10 +341,10 @@ void ManagedScan::calcReducedShow()
     oct->GetOctTreeCenter(center);
   }
 
-  unsigned int size = center.size();
+  size_t size = center.size();
   TripleArray<float> xyz_r(m_shared_scan->createXYZReducedShow(size));
-  for(unsigned int i = 0; i < size; ++i) {
-    for(unsigned int j = 0; j < 3; ++j) {
+  for(size_t i = 0; i < size; ++i) {
+    for(size_t j = 0; j < 3; ++j) {
       xyz_r[i][j] = center[i][j];
     }
   }
@@ -374,13 +374,13 @@ void ManagedScan::createOcttree()
                                   true);
     } else { // without reduction, xyz + attribute points
       float** pts = octtree_pointtype.createPointArray<float>(this);
-      unsigned int nrpts = size<DataXYZ>("xyz");
+      size_t nrpts = size<DataXYZ>("xyz");
       btree = new BOctTree<float>(pts,
                                   nrpts,
                                   octtree_voxelSize,
                                   octtree_pointtype,
                                   true);
-      for(unsigned int i = 0; i < nrpts; ++i) delete[] pts[i]; delete[] pts;
+      for(size_t i = 0; i < nrpts; ++i) delete[] pts[i]; delete[] pts;
     }
     // save created octtree
     if(octtree_saveOct) {
@@ -391,7 +391,7 @@ void ManagedScan::createOcttree()
 
   // copy tree into cache
   try {
-    unsigned int size = btree->getMemorySize();
+    size_t size = btree->getMemorySize();
     unsigned char* mem_ptr
       = m_shared_scan->createOcttree(size).get_raw_pointer();
     new(mem_ptr) BOctTree<float>(*btree, mem_ptr, size);
@@ -404,7 +404,7 @@ void ManagedScan::createOcttree()
 }
 
 
-unsigned int ManagedScan::readFrames()
+size_t ManagedScan::readFrames()
 {
   // automatically read on getFrames
   return m_shared_scan->getFrames().size();
@@ -415,12 +415,12 @@ void ManagedScan::saveFrames()
   m_shared_scan->saveFrames();
 }
 
-unsigned int ManagedScan::getFrameCount()
+size_t ManagedScan::getFrameCount()
 {
   return m_shared_scan->getFrames().size();
 }
 
-void ManagedScan::getFrame(unsigned int i,
+void ManagedScan::getFrame(size_t i,
                            const double*& pose_matrix,
                            AlgoType& type)
 {
